Major update for v5

+++ web/Groups.pl
@@ -4,17 +4,21 @@ use strict;
 use CGI;
 #use CGI::Ajax;
 use JSON::XS;
-use CoGe::Accessory::LogUser;
-use CoGe::Accessory::Web;
-use CoGeX;
+use lib '/home/mbomhoff/CoGe/Accessory/lib'; #FIXME remove
+use lib '/home/mbomhoff/CoGeX/lib'; #FIXME 8/2/12 remove
+use CoGe_dev::Accessory::LogUser;
+use CoGe_dev::Accessory::Web;
+use CoGeX_dev;
 use HTML::Template;
 use Digest::MD5 qw(md5_base64);
 use URI::Escape;
 use Data::Dumper;
+use File::Path;
+
 no warnings 'redefine';
 
 use vars qw($P $DBNAME $DBHOST $DBPORT $DBUSER $DBPASS $connstr $PAGE_NAME $TEMPDIR $USER $DATE $BASEFILE $coge $cogeweb %FUNCTION $COOKIE_NAME $FORM $URL $COGEDIR $TEMPDIR $TEMPURL);
-$P = CoGe::Accessory::Web::get_defaults($ENV{HOME}.'coge.conf');
+$P = CoGe_dev::Accessory::Web::get_defaults($ENV{HOME}.'coge.conf');
 
 $DATE = sprintf( "%04d-%02d-%02d %02d:%02d:%02d",
 		sub { ($_[5]+1900, $_[4]+1, $_[3]),$_[2],$_[1],$_[0] }->(localtime));
@@ -27,20 +31,20 @@ $DBPORT = $P->{DBPORT};
 $DBUSER = $P->{DBUSER};
 $DBPASS = $P->{DBPASS};
 $connstr = "dbi:mysql:dbname=".$DBNAME.";host=".$DBHOST.";port=".$DBPORT;
-$coge = CoGeX->connect($connstr, $DBUSER, $DBPASS );
+$coge = CoGeX_dev->connect($connstr, $DBUSER, $DBPASS );
 
 $COOKIE_NAME = $P->{COOKIE_NAME};
 $URL = $P->{URL};
 $COGEDIR = $P->{COGEDIR};
-$TEMPDIR = $P->{TEMPDIR}."GenomeList/";
+$TEMPDIR = $P->{TEMPDIR}."Groups/";
 mkpath ($TEMPDIR, 0,0777) unless -d $TEMPDIR;
-$TEMPURL = $P->{TEMPURL}."GenomeList/";
+$TEMPURL = $P->{TEMPURL}."Groups/";
 
 
 my ($cas_ticket) =$FORM->param('ticket');
 $USER = undef;
-($USER) = CoGe::Accessory::Web->login_cas(cookie_name=>$COOKIE_NAME, ticket=>$cas_ticket, coge=>$coge, this_url=>$FORM->url()) if($cas_ticket);
-($USER) = CoGe::Accessory::LogUser->get_user(cookie_name=>$COOKIE_NAME,coge=>$coge) unless $USER;
+($USER) = CoGe_dev::Accessory::Web->login_cas(cookie_name=>$COOKIE_NAME, ticket=>$cas_ticket, coge=>$coge, this_url=>$FORM->url()) if($cas_ticket);
+($USER) = CoGe_dev::Accessory::LogUser->get_user(cookie_name=>$COOKIE_NAME,coge=>$coge) unless $USER;
 
 %FUNCTION = (
 	     gen_html=>\&gen_html,
@@ -48,12 +52,14 @@ $USER = undef;
 	     create_group=>\&create_group,
 	     delete_group=>\&delete_group,
 	     update_group=>\&update_group,
-	     get_group_info=>\&get_group_info,
+	     modify_users=>\&modify_users,
+	     edit_group_info=>\&edit_group_info,
 	     add_user_to_group=>\&add_user_to_group,
 	     remove_user_from_group=>\&remove_user_from_group,
+	     add_list_to_group=>\&add_list_to_group,
+	     remove_list_from_group=>\&remove_list_from_group,
 	     add_genome_to_group=>\&add_genome_to_group,
 	     remove_genome_from_group=>\&remove_genome_from_group,
-	     get_orphan_genomes=>\&get_orphan_genomes,
     );
 
 dispatch();
@@ -119,27 +125,6 @@ sub gen_body
       return $template->output;
   }
 
-sub get_orphan_genomes
-  {
-    my %opts = @_;
-    return "Naughty monkey!" unless $USER->is_admin;
-    my @genomes;
-    #time to deal with admin privileges
-    foreach my $genome (sort {$a->organism->name cmp $b->organism->name} $coge->resultset('DatasetGroup')->search({restricted=>1}))
-      {
-	my $name = $genome->organism->name;
-	$name .= " ".join (", ", map{$_->name} $genome->source) .": ";
-	$name .= $genome->name.", " if $genome->name;
-	$name .=  "v".$genome->version." ".$genome->type->name;#." ".commify($genome->length)."nt";
-	#my $name = $genome->organism->name;
-	#$name .= " (v".$genome->version.")";
-	$name = qq{<span class="link" onclick="window.open('OrganismView.pl?dsgid=}.$genome->id.qq{');">}.$name."</span>";
-	push @genomes, $name;
-      }
-    my $output = join ("<br>",@genomes);
-    return $output;
-  }
-
 sub get_roles
   {
     my @roles;
@@ -160,6 +145,13 @@ sub add_user_to_group
     my $uid = $opts{uid};
 #    return 1 if $uid == $USER->id;
     return "UGID and/or UID not specified"  unless $ugid && $uid;
+    my $group = $coge->resultset('UserGroup')->find($ugid);
+    if ($group->locked && !$USER->is_admin)
+      {
+	return "This is a locked group.  Admin permission is needed to modify.";
+      }
+    my @res = $coge->resultset('UserGroupConnector')->search({user_id=>$uid, user_group_id=>$ugid});
+    return 1 if @res; #previously added;
     my ($ugc) = $coge->resultset('UserGroupConnector')->create({user_id=>$uid, user_group_id=>$ugid});
     return 1;
   }
@@ -171,8 +163,16 @@ sub remove_user_from_group
     my $uid = $opts{uid};
     if ($uid == $USER->id && !$USER->is_admin) #only allow this for admins
       { return "Can't remove yourself from this group!"; }
-    my ($ugc) = $coge->resultset('UserGroupConnector')->search({user_id=>$uid, user_group_id=>$ugid});
-    $ugc->delete;
+    my $group = $coge->resultset('UserGroup')->find($ugid);
+    if ($group->locked && !$USER->is_admin)
+      {
+	return "This is a locked group.  Admin permission is needed to modify.";
+      }
+
+    foreach my $ugc ($coge->resultset('UserGroupConnector')->search({user_id=>$uid, user_group_id=>$ugid}))
+      {
+	$ugc->delete;
+      }
     return 1;
   }
 sub add_genome_to_group
@@ -184,7 +184,7 @@ sub add_genome_to_group
     my $is_owner = $USER->is_owner(dsg=>$dsgid);
     $is_owner=1 if $USER->is_admin;
     return "You are not the owner of this genome" unless $is_owner;
-    my ($ugdc) = $coge->resultset('UserGroupDataConnector')->create({dataset_group_id=>$dsgid, user_group_id=>$ugid});
+    my ($ugdc) = $coge->resultset('UserGroupDataConnector')->create({genome_id=>$dsgid, user_group_id=>$ugid});
     return 1;
   }
 
@@ -194,23 +194,21 @@ sub remove_genome_from_group
     my $ugid = $opts{ugid};
     my $dsgid = $opts{dsgid};
     return "DSGID and/or UGID not specified" unless $dsgid && $ugid;
-    my ($ugdc) = $coge->resultset('UserGroupDataConnector')->search({dataset_group_id=>$dsgid, user_group_id=>$ugid});
+    my ($ugdc) = $coge->resultset('UserGroupDataConnector')->search({genome_id=>$dsgid, user_group_id=>$ugid});
     $ugdc->delete;
     return 1;
   }
 
-sub get_group_info
+
+sub modify_users
   {
     my %opts = @_;
     my $ugid = $opts{ugid};
     return 0 unless $ugid;
     my %data;
     my $group = $coge->resultset('UserGroup')->find($ugid);
-    $data{title} = $group->name;
+    $data{title} = $group->name."(id".$group->id.")";
     $data{title} .= ": ".$group->description if $group->description;
-    $data{name} = $group->name;
-    $data{desc} ="";
-    $data{desc} = $group->description if $group->description;
     my %users;
     my @users;
     foreach my $user (sort{$a->last_name cmp $b->last_name || $a->user_name cmp $b->user_name} $group->users)
@@ -234,49 +232,49 @@ sub get_group_info
       }
     $data{all_users}=\@all_users;
     
-    my @genomes;
-    my %genomes;
-    foreach my $genome ($group->private_genomes)
-      {
-	my $name = $genome->organism->name;
-	$name .= " ".join (", ", map{$_->name} $genome->source) .": ";
-        $name .= $genome->name.", " if $genome->name;
-        $name .=  "v".$genome->version." ".$genome->type->name;#." ".commify($genome->length)."nt";
-	push @genomes, {dsgid_name=>$name, dsgid=>$genome->id};
-	$genomes{$genome->id}=1;
-      }
-    $data{genomes}=\@genomes;
+ #   my @genomes;
+ #   my %genomes;
+ #   foreach my $genome ($group->private_genomes)
+ #     {
+#	my $name = $genome->organism->name;
+#	$name .= " ".join (", ", map{$_->name} $genome->source) .": ";
+ #       $name .= $genome->name.", " if $genome->name;
+#        $name .=  "v".$genome->version." ".$genome->type->name;#." ".commify($genome->length)."nt";
+#	push @genomes, {dsgid_name=>$name, dsgid=>$genome->id};
+#	$genomes{$genome->id}=1;
+ #     }
+ #   $data{genomes}=\@genomes;
 
-    my @genome_list;
+ #   my @genome_list;
     #time to deal with admin privileges
-    if ($USER->is_admin)
-      {
-	@genome_list = $coge->resultset('DatasetGroup')->search({restricted=>1});
-      }
-    else
-      {
-	@genome_list = $USER->private_genomes;
-      }
-    my @all_genomes;
-    foreach my $genome (sort {$a->organism->name cmp $b->organism->name} @genome_list)
-      {
-	next if $genomes{$genome->id}; #skip users we already have
-	my $name = $genome->organism->name;
-	$name .= " ".join (", ", map{$_->name} $genome->source) .": ";
-	$name .= $genome->name.", " if $genome->name;
-	$name .=  "v".$genome->version." ".$genome->type->name;#." ".commify($genome->length)."nt";
-	push @all_genomes, {dsgid_name=>$name, dsgid=>$genome->id};
-      }
-    $data{all_genomes}=\@all_genomes;
+ #   if ($USER->is_admin)
+ #     {
+#	@genome_list = $coge->resultset('Genome')->search({restricted=>1});
+ #     }
+ #   else
+ #     {
+#	@genome_list = $USER->private_genomes;
+#      }
+#    my @all_genomes;
+#    foreach my $genome (sort {$a->organism->name cmp $b->organism->name} @genome_list)
+#      {
+#	next if $genomes{$genome->id}; #skip users we already have
+#	my $name = $genome->organism->name;
+#	$name .= " ".join (", ", map{$_->name} $genome->source) .": ";
+#	$name .= $genome->name.", " if $genome->name;
+#	$name .=  "v".$genome->version." ".$genome->type->name;#." ".commify($genome->length)."nt";
+#	push @all_genomes, {dsgid_name=>$name, dsgid=>$genome->id};
+#      }
+#    $data{all_genomes}=\@all_genomes;
     my $template = HTML::Template->new(filename=>$P->{TMPLDIR}.'Groups.tmpl');
-    $template->param(EDIT_GROUP=>1);
+    $template->param(MODIFY_USERS=>1);
     $template->param(UGID=>$ugid);
-    $template->param(NAME=>$data{name});
-    $template->param(DESC=>$data{desc});
+#    $template->param(NAME=>$data{name});
+#    $template->param(DESC=>$data{desc});
     $template->param(UGID_LOOP=>$data{users});
     $template->param(ALL_UGID_LOOP=>$data{all_users});
-    $template->param(DSGID_LOOP=>$data{genomes});
-    $template->param(ALL_DSGID_LOOP=>$data{all_genomes});
+#    $template->param(DSGID_LOOP=>$data{genomes});
+#    $template->param(ALL_DSGID_LOOP=>$data{all_genomes});
     #add message if user is an admin
     if ($USER->is_admin)
       {
@@ -286,6 +284,26 @@ sub get_group_info
     return encode_json (\%data);
   }
 
+sub edit_group_info
+  {
+    my %opts = @_;
+    my $ugid = $opts{ugid};
+    return 0 unless $ugid;
+    my %data;
+    my $group = $coge->resultset('UserGroup')->find($ugid);
+    $data{title} = $group->name."(id".$group->id.")";
+    $data{title} .= ": ".$group->description if $group->description;
+    $data{name} = $group->name;
+    $data{desc} ="";
+    $data{desc} = $group->description if $group->description;
+    my $template = HTML::Template->new(filename=>$P->{TMPLDIR}.'Groups.tmpl');
+    $template->param(EDIT_GROUP_INFO=>1);
+    $template->param(UGID=>$ugid);
+    $template->param(NAME=>$data{name});
+    $template->param(DESC=>$data{desc});    $data{output}=$template->output;
+    return encode_json (\%data);
+  }
+ 
 sub create_group
   {
     my %opts = @_;
@@ -321,6 +339,11 @@ sub delete_group
     my %opts = @_;
     my $ugid = $opts{ugid};
     my $group = $coge->resultset('UserGroup')->find($ugid);
+    return unless $group;
+    if ($group->locked && !$USER->is_admin)
+      {
+	return "This is a locked group.  Admin permission is needed to delete.";
+      }
     $group->delete();
   }
 
@@ -340,13 +363,13 @@ sub get_groups_for_user
     foreach my $group (@group_list)
       {
         my %groups;
-        $groups{NAME}=$group->name;
+        $groups{NAME}=$group->name." (id".$group->id.")";
         $groups{DESC}=$group->description if $group->description;
         my $role = $group->role->name;
 #        $role .= ": ".$group->role->description if $group->role->description;
         $groups{ROLE}=$role;
-        my $perm = join (", ", map {$_->name} $group->role->permissions);
-        $groups{PERM}=$perm;
+#        my $perm = join (", ", map {$_->name} $group->role->permissions);
+#        $groups{PERM}=$perm;
 	my @users;
 	foreach my $user (sort {$a->last_name cmp $b->last_name} $group->users)
 	  {
@@ -358,17 +381,35 @@ sub get_groups_for_user
 	  }
 #	push @users, "Self only" unless @users;
 	$groups{MEM}=join (",<br>", @users);
-	
-	my @genome;
-	push @genome, "Apotheosis" if $group->role->name =~ /admin/i;
-	foreach my $genome (sort {$a->organism->name cmp $b->organism->name || $a->version cmp $b->version || $a->genomic_sequence_type_id  <=> $b->genomic_sequence_type_id} $group->private_genomes)
+	my %lists;
+	foreach my $list (sort {$a->type->name cmp $b->type->name || $a->name cmp $b->name} $group->lists)
 	  {
-	    my $name = $genome->organism->name;
-	    $name .= " (v".$genome->version.", ".$genome->type->name.")";
-	    $name = qq{<span class="link" onclick="window.open('OrganismView.pl?dsgid=}.$genome->id.qq{');">}.$name."</span>";
-	    push @genome, $name;
+	    my $name;
+	    $name .= "(P)" if !$list->restricted; #list is open to the public (P)
+	    $name .= $list->name;
+	    $name = qq{<span class="link" onclick="window.open('ListView.pl?lid=}.$list->id.qq{');">}.$name."</span>";
+	    push @{$lists{$list->type->name}}, $name;
 	  }
-	$groups{GENOMES}=join (",<br>", @genome);
+	my $lists = "<table class='small'>";
+	foreach my $type (sort keys %lists)
+	  {
+	    $lists .= qq{<tr><td>$type<td>}.join ("<br>", @{$lists{$type}});
+	  }
+	$lists.="</table>";
+	$groups{LISTS}=join (",<br>", $lists);
+	
+###	my @genome;
+#	push @genome, "Apotheosis" if $group->role->name =~ /admin/i;
+#	foreach my $genome (sort {$a->organism->name cmp $b->organism->name || $a->version cmp $b->version || $a->genomic_sequence_type_id  <=> $b->genomic_sequence_type_id} $group->genomes)
+#	  {
+#	    my $name = $genome->organism->name;
+#	    $name .= " (v".$genome->version.", ".$genome->type->name.")";
+#	    $name = qq{<span class="link" onclick="window.open('OrganismView.pl?dsgid=}.$genome->id.qq{');">}.$name."</span>";
+#	    push @genome, $name;
+#	  }
+#	$groups{GENOMES}=join (",<br>", @genome);
+
+
 	$groups{UGID}=$group->id;
         push @groups, \%groups;
       }
